+++ app/QuestionChoice.php
@@ -2,14 +2,17 @@
 
 namespace Gamify;
 
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 
 /**
  * Class QuestionChoice.
  *
- * @property  string $text    The text of this choice.
- * @property  bool   $correct Is this choice correct?.
- * @property  int    $score   How many points are added by this choice.
+ * @property string                $text    The text of this choice.
+ * @property int                   $score   How many points are added by this choice.
+ * @mixin \Eloquent
+ * @property-read \Gamify\Question $question
+ * @method static \Illuminate\Database\Eloquent\Builder|\Gamify\QuestionChoice correct()
  */
 class QuestionChoice extends Model
 {
@@ -34,7 +37,6 @@ class QuestionChoice extends Model
      */
     protected $fillable = [
         'text',
-        'correct',
         'score',
     ];
 
@@ -46,7 +48,6 @@ class QuestionChoice extends Model
     protected $casts = [
         'id' => 'int',
         'text' => 'string',
-        'correct' => 'bool',
         'score' => 'int',
     ];
 
@@ -64,4 +65,37 @@ class QuestionChoice extends Model
      * Every time we modify a choice we need to touch the question.
      */
     protected $touches = ['question'];
+
+    /**
+     * Returns true if the choice is considered correct.
+     *
+     * @return bool
+     */
+    public function isCorrect(): bool
+    {
+        return $this->score > 0;
+    }
+
+    /**
+     * Return question choices considered as correct.
+     *
+     * @param \Illuminate\Database\Eloquent\Builder $query
+     *
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public function scopeCorrect(Builder $query): Builder
+    {
+        return $query->where('score', '>', '0');
+    }
+
+    /**
+     * DEPRECATED: Use isCorrect() instead.
+     *
+     * @return bool
+     * @deprecated
+     */
+    public function getCorrectAttribute(): bool
+    {
+        return $this->isCorrect();
+    }
 }